[PAGOPA-1836] fix: moved station check on RT sending during RT generation
